requirements.txt
asgiref==3.8.1
attrs==25.3.0
dj-database-url==2.3.0
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.3.1
gunicorn==23.0.0
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
npm==0.1.1
optional-django==0.1.0
packaging==24.2
psycopg==3.2.6
psycopg-binary==3.2.6
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.23.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
whitenoise==6.9.0
